Deduplicate stream-required error detection
